requirements.txt
torch
openai-whisper
transformers
websockets
asyncio
